backend-proxy-render/requirements.txt
fastapi
uvicorn
# Integrated TLS Impersonation via curl-cffi
yt-dlp[default,curl-cffi]
httpx
python-dotenv
pytubefix
